🧹 Extract scroll listener to useScroll hook
